+++ src/main/java/se/sundsvall/casedata/integration/db/AttachmentRepository.java
@@ -1,15 +1,13 @@
 package se.sundsvall.casedata.integration.db;
 
-import java.util.List;
-import java.util.Optional;
-
+import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
 import org.javers.spring.annotation.JaversSpringDataAuditable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-
 import se.sundsvall.casedata.integration.db.model.AttachmentEntity;
 
-import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
+import java.util.List;
+import java.util.Optional;
 
 @JaversSpringDataAuditable
 @CircuitBreaker(name = "attachmentRepository")